+++ dart/geobase/lib/src/vector_data/model.dart
@@ -1,0 +1,2 @@
+export 'model/bounded.dart';
+export 'model/geometry.dart';