+++ dart/geobase/lib/src/vector_data/model/geometry.dart
@@ -1,0 +1,9 @@
+export 'geometry/geometry.dart';
+export 'geometry/geometry_builder.dart';
+export 'geometry/geometry_collection.dart';
+export 'geometry/linestring.dart';
+export 'geometry/multi_linestring.dart';
+export 'geometry/multi_point.dart';
+export 'geometry/multi_polygon.dart';
+export 'geometry/point.dart';
+export 'geometry/polygon.dart';