+++ dart/geobase/CHANGELOG.md
@@ -20,6 +20,8 @@ binary data formats, encodings and content interfaces also redesigned.
 - [Coordinate order on box and position consistently #134](https://github.com/navibyte/geospatial/issues/134)
 - [Positions as iterable of coordinate values on geobase #136](https://github.com/navibyte/geospatial/issues/136)
 - [PositionData access (on lists or buffers) #137](https://github.com/navibyte/geospatial/issues/137)
+- new mini library "transforms" and "projections"
+- [Simple geometry data structures on geobase #133](https://github.com/navibyte/geospatial/issues/133)
 
 🛠 Refactoring:
 - [Optimize reading from iterable of coordinate values #131](https://github.com/navibyte/geospatial/issues/131)